# _openapi.json
{"openapi":"3.1.0","info":{"title":"PYTHONPathFinder","version":"2"},"paths":{"/api/documents/open":{"post":{"tags":["documents"],"summary":"Open Document","operationId":"open_document_api_documents_open_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentReference"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentView"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/preview-csv":{"post":{"tags":["documents"],"summary":"Preview Csv","operationId":"preview_csv_api_documents_preview_csv_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CsvPreviewRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CsvPreviewView"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/translations/batch":{"post":{"tags":["translations"],"summary":"Translate Batch","operationId":"translate_batch_api_translations_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchTranslationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchTranslationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/changes/preview":{"post":{"tags":["changes"],"summary":"Preview Changes","operationId":"preview_changes_api_changes_preview_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeBatch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePreviewView"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/changes/apply":{"post":{"tags":["changes"],"summary":"Apply Changes","operationId":"apply_changes_api_changes_apply_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeBatch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeResultView"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workspace/project":{"post":{"tags":["workspace"],"summary":"Project Workspace","operationId":"project_workspace_api_workspace_project_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkspaceProjectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkspaceProjectionView"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workspace/policy":{"get":{"tags":["workspace"],"summary":"Workspace Upload Policy","operationId":"workspace_upload_policy_api_workspace_policy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkspaceUploadPolicyView"}}}}}}},"/api/workspace/files":{"get":{"tags":["workspace"],"summary":"List Workspace Files","operationId":"list_workspace_files_api_workspace_files_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WorkspaceFileView"},"type":"array","title":"Response List Workspace Files Api Workspace Files Get"}}}}}},"post":{"tags":["workspace"],"summary":"Upload Workspace Files","operationId":"upload_workspace_files_api_workspace_files_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_workspace_files_api_workspace_files_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WorkspaceFileView"},"type":"array","title":"Response Upload Workspace Files Api Workspace Files Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workspace/download/{file_path}":{"get":{"tags":["workspace"],"summary":"Download Workspace File","operationId":"download_workspace_file_api_workspace_download__file_path__get","parameters":[{"name":"file_path","in":"path","required":true,"schema":{"type":"string","title":"File Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workspace/archive":{"get":{"tags":["workspace"],"summary":"Download Workspace Archive","operationId":"download_workspace_archive_api_workspace_archive_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/sql/inspect":{"post":{"tags":["sql"],"summary":"Inspect Sql","operationId":"inspect_sql_api_sql_inspect_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SqlModelRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SqlModelView"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sql/apply-action":{"post":{"tags":["sql"],"summary":"Apply Sql Action","operationId":"apply_sql_action_api_sql_apply_action_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SqlActionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SqlActionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Api Health Get"}}}}}}}},"components":{"schemas":{"ArtifactView":{"properties":{"id":{"type":"string","title":"Id"},"path":{"type":"string","title":"Path"},"label":{"type":"string","title":"Label"},"conditional":{"type":"boolean","title":"Conditional","default":false},"in_loop":{"type":"boolean","title":"In Loop","default":false},"producer_step_ids":{"items":{"type":"string"},"type":"array","title":"Producer Step Ids"},"consumer_step_ids":{"items":{"type":"string"},"type":"array","title":"Consumer Step Ids"},"order_valid":{"type":"boolean","title":"Order Valid","default":true},"is_external_input":{"type":"boolean","title":"Is External Input","default":false},"is_output":{"type":"boolean","title":"Is Output","default":false}},"type":"object","required":["id","path","label"],"title":"ArtifactView"},"BatchTranslationRequest":{"properties":{"source_paths":{"items":{"type":"string"},"type":"array","maxItems":100,"minItems":1,"title":"Source Paths"},"out_dir":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Out Dir"}},"type":"object","required":["source_paths"],"title":"BatchTranslationRequest"},"BatchTranslationResponse":{"properties":{"documents":{"items":{"$ref":"#/components/schemas/DocumentView"},"type":"array","title":"Documents"},"diagnostics":{"items":{"$ref":"#/components/schemas/DiagnosticView"},"type":"array","title":"Diagnostics"}},"type":"object","required":["documents","diagnostics"],"title":"BatchTranslationResponse"},"Body_upload_workspace_files_api_workspace_files_post":{"properties":{"files":{"items":{"type":"string","contentMediaType":"application/octet-stream"},"type":"array","title":"Files"},"paths":{"items":{"type":"string"},"type":"array","title":"Paths"}},"type":"object","required":["files","paths"],"title":"Body_upload_workspace_files_api_workspace_files_post"},"ChangeBatch":{"properties":{"source_path":{"type":"string","title":"Source Path"},"output_path":{"type":"string","title":"Output Path"},"source_hash":{"type":"string","title":"Source Hash"},"output_hash":{"type":"string","title":"Output Hash"},"revision":{"type":"integer","title":"Revision"},"changes":{"items":{"$ref":"#/components/schemas/ParameterChangeRequest"},"type":"array","minItems":1,"title":"Changes"}},"type":"object","required":["source_path","output_path","source_hash","output_hash","revision","changes"],"title":"ChangeBatch"},"ChangePreviewView":{"properties":{"valid":{"type":"boolean","title":"Valid"},"diff":{"type":"string","title":"Diff"},"issues":{"items":{"$ref":"#/components/schemas/ValidationIssueView"},"type":"array","title":"Issues"}},"type":"object","required":["valid","diff"],"title":"ChangePreviewView"},"ChangeResultView":{"properties":{"document":{"$ref":"#/components/schemas/DocumentView"}},"type":"object","required":["document"],"title":"ChangeResultView"},"CsvPreviewRequest":{"properties":{"source_path":{"type":"string","title":"Source Path"},"csv_path":{"type":"string","title":"Csv Path"}},"type":"object","required":["source_path","csv_path"],"title":"CsvPreviewRequest"},"CsvPreviewView":{"properties":{"path":{"type":"string","title":"Path"},"columns":{"items":{"type":"string"},"type":"array","title":"Columns"},"rows":{"items":{"items":{"type":"string"},"type":"array"},"type":"array","title":"Rows"},"truncated":{"type":"boolean","title":"Truncated"},"size_bytes":{"type":"integer","title":"Size Bytes"}},"type":"object","required":["path","columns","rows","truncated","size_bytes"],"title":"CsvPreviewView"},"DependencyIssueView":{"properties":{"code":{"type":"string","enum":["BROKEN_DEPENDENCY","DUPLICATE_OUTPUT"],"title":"Code"},"document_id":{"type":"string","title":"Document Id"},"step_id":{"type":"string","title":"Step Id"},"artifact":{"type":"string","title":"Artifact"},"message":{"type":"string","title":"Message"},"related_document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Related Document Id"},"related_step_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Related Step Id"}},"type":"object","required":["code","document_id","step_id","artifact","message"],"title":"DependencyIssueView"},"DependencyLinkView":{"properties":{"artifact":{"type":"string","title":"Artifact"},"producer_document_id":{"type":"string","title":"Producer Document Id"},"producer_step_id":{"type":"string","title":"Producer Step Id"},"consumer_document_id":{"type":"string","title":"Consumer Document Id"},"consumer_step_id":{"type":"string","title":"Consumer Step Id"}},"type":"object","required":["artifact","producer_document_id","producer_step_id","consumer_document_id","consumer_step_id"],"title":"DependencyLinkView"},"DiagnosticView":{"properties":{"level":{"type":"string","enum":["info","warning","error"],"title":"Level"},"code":{"type":"string","title":"Code"},"message":{"type":"string","title":"Message"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"node_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Node Id"}},"type":"object","required":["level","code","message"],"title":"DiagnosticView"},"DocumentReference":{"properties":{"source_path":{"type":"string","title":"Source Path"},"output_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output Path"}},"type":"object","required":["source_path"],"title":"DocumentReference"},"DocumentView":{"properties":{"schema_version":{"type":"integer","title":"Schema Version","default":2},"id":{"type":"string","title":"Id"},"source_path":{"type":"string","title":"Source Path"},"output_path":{"type":"string","title":"Output Path"},"source_hash":{"type":"string","title":"Source Hash"},"output_hash":{"type":"string","title":"Output Hash"},"revision":{"type":"integer","title":"Revision","default":1},"synchronized":{"type":"boolean","title":"Synchronized","default":true},"read_only_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Read Only Reason"},"steps":{"items":{"$ref":"#/components/schemas/StepView"},"type":"array","title":"Steps"},"scopes":{"items":{"$ref":"#/components/schemas/ScopeView"},"type":"array","title":"Scopes"},"artifacts":{"items":{"$ref":"#/components/schemas/ArtifactView"},"type":"array","title":"Artifacts"},"diagnostics":{"items":{"$ref":"#/components/schemas/DiagnosticView"},"type":"array","title":"Diagnostics"}},"type":"object","required":["id","source_path","output_path","source_hash","output_hash","steps","scopes","artifacts","diagnostics"],"title":"DocumentView"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ParameterChangeRequest":{"properties":{"parameter_id":{"type":"string","title":"Parameter Id"},"value":{"title":"Value"}},"type":"object","required":["parameter_id","value"],"title":"ParameterChangeRequest"},"ParameterView":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"position":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Position"},"source":{"type":"string","title":"Source"},"value":{"title":"Value"},"editor_type":{"type":"string","enum":["string","multiline","integer","boolean","list","dynamic"],"title":"Editor Type"},"editable":{"type":"boolean","title":"Editable"},"read_only_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Read Only Reason"},"constraints":{"additionalProperties":true,"type":"object","title":"Constraints"},"annotation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Annotation"},"required":{"type":"boolean","title":"Required","default":true},"default":{"title":"Default"},"capabilities":{"items":{"type":"string"},"type":"array","title":"Capabilities"}},"type":"object","required":["id","name","source","editor_type","editable"],"title":"ParameterView"},"ProjectedDocumentView":{"properties":{"document_id":{"type":"string","title":"Document Id"},"artifacts":{"items":{"$ref":"#/components/schemas/ArtifactView"},"type":"array","title":"Artifacts"}},"type":"object","required":["document_id","artifacts"],"title":"ProjectedDocumentView"},"ScopeView":{"properties":{"id":{"type":"string","title":"Id"},"node_kind":{"type":"string","enum":["if","branch","loop"],"title":"Node Kind"},"scope_kind":{"type":"string","title":"Scope Kind"},"label":{"type":"string","title":"Label"},"start_index":{"type":"integer","title":"Start Index"},"end_index":{"type":"integer","title":"End Index"},"parent_scope_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Scope Id"}},"type":"object","required":["id","node_kind","scope_kind","label","start_index","end_index"],"title":"ScopeView"},"SourceSpanView":{"properties":{"file":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"File"},"start_line":{"type":"integer","title":"Start Line"},"end_line":{"type":"integer","title":"End Line"}},"type":"object","required":["start_line","end_line"],"title":"SourceSpanView"},"SqlActionRequest":{"properties":{"source_path":{"type":"string","title":"Source Path"},"output_path":{"type":"string","title":"Output Path"},"parameter_id":{"type":"string","title":"Parameter Id"},"changes":{"items":{"$ref":"#/components/schemas/ParameterChangeRequest"},"type":"array","title":"Changes"},"action":{"type":"string","title":"Action"},"arguments":{"additionalProperties":true,"type":"object","title":"Arguments"}},"type":"object","required":["source_path","output_path","parameter_id","action"],"title":"SqlActionRequest"},"SqlActionResponse":{"properties":{"change":{"$ref":"#/components/schemas/ParameterChangeRequest"},"model":{"$ref":"#/components/schemas/SqlModelView"}},"type":"object","required":["change","model"],"title":"SqlActionResponse"},"SqlEditCapabilitiesView":{"properties":{"selected":{"type":"boolean","title":"Selected"},"filters":{"type":"boolean","title":"Filters"},"joins":{"type":"boolean","title":"Joins"},"raw_sql":{"type":"boolean","title":"Raw Sql","default":true}},"type":"object","required":["selected","filters","joins"],"title":"SqlEditCapabilitiesView"},"SqlJoinView":{"properties":{"id":{"type":"string","title":"Id"},"join_type":{"type":"string","title":"Join Type"},"source":{"type":"string","title":"Source"},"predicates":{"items":{"$ref":"#/components/schemas/SqlPredicateView"},"type":"array","title":"Predicates"},"editable_type":{"type":"boolean","title":"Editable Type"},"editable_source":{"type":"boolean","title":"Editable Source"},"read_only_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Read Only Reason"},"span":{"$ref":"#/components/schemas/SqlSpanView"},"type_span":{"$ref":"#/components/schemas/SqlSpanView"},"source_span":{"$ref":"#/components/schemas/SqlSpanView"}},"type":"object","required":["id","join_type","source","predicates","editable_type","editable_source","span","type_span","source_span"],"title":"SqlJoinView"},"SqlModelRequest":{"properties":{"source_path":{"type":"string","title":"Source Path"},"output_path":{"type":"string","title":"Output Path"},"parameter_id":{"type":"string","title":"Parameter Id"},"changes":{"items":{"$ref":"#/components/schemas/ParameterChangeRequest"},"type":"array","title":"Changes"}},"type":"object","required":["source_path","output_path","parameter_id"],"title":"SqlModelRequest"},"SqlModelView":{"properties":{"source":{"type":"string","title":"Source"},"filter_operators":{"items":{"type":"string"},"type":"array","title":"Filter Operators"},"join_types":{"items":{"type":"string"},"type":"array","title":"Join Types"},"logical_connectors":{"items":{"type":"string"},"type":"array","title":"Logical Connectors"},"statement_span":{"$ref":"#/components/schemas/SqlSpanView"},"selections":{"items":{"$ref":"#/components/schemas/SqlSelectionView"},"type":"array","title":"Selections"},"filters":{"items":{"$ref":"#/components/schemas/SqlPredicateView"},"type":"array","title":"Filters"},"joins":{"items":{"$ref":"#/components/schemas/SqlJoinView"},"type":"array","title":"Joins"},"sources":{"items":{"$ref":"#/components/schemas/SqlSourceView"},"type":"array","title":"Sources"},"capabilities":{"$ref":"#/components/schemas/SqlEditCapabilitiesView"},"read_only_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Read Only Reason"},"select_list_span":{"anyOf":[{"$ref":"#/components/schemas/SqlSpanView"},{"type":"null"}]},"where_clause_span":{"anyOf":[{"$ref":"#/components/schemas/SqlSpanView"},{"type":"null"}]},"where_body_span":{"anyOf":[{"$ref":"#/components/schemas/SqlSpanView"},{"type":"null"}]},"from_clause_span":{"anyOf":[{"$ref":"#/components/schemas/SqlSpanView"},{"type":"null"}]}},"type":"object","required":["source","filter_operators","join_types","logical_connectors","statement_span","selections","filters","joins","sources","capabilities"],"title":"SqlModelView"},"SqlPredicateView":{"properties":{"id":{"type":"string","title":"Id"},"left":{"type":"string","title":"Left"},"operator":{"type":"string","title":"Operator"},"right":{"type":"string","title":"Right"},"connector":{"anyOf":[{"type":"string","enum":["AND","OR"]},{"type":"null"}],"title":"Connector"},"raw":{"type":"string","title":"Raw"},"editable":{"type":"boolean","title":"Editable"},"read_only_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Read Only Reason"},"span":{"$ref":"#/components/schemas/SqlSpanView"},"connector_span":{"anyOf":[{"$ref":"#/components/schemas/SqlSpanView"},{"type":"null"}]}},"type":"object","required":["id","left","operator","right","raw","editable","span"],"title":"SqlPredicateView"},"SqlSelectionView":{"properties":{"id":{"type":"string","title":"Id"},"expression":{"type":"string","title":"Expression"},"alias":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Alias"},"raw":{"type":"string","title":"Raw"},"editable":{"type":"boolean","title":"Editable"},"read_only_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Read Only Reason"},"span":{"$ref":"#/components/schemas/SqlSpanView"}},"type":"object","required":["id","expression","raw","editable","span"],"title":"SqlSelectionView"},"SqlSourceView":{"properties":{"id":{"type":"string","title":"Id"},"expression":{"type":"string","title":"Expression"},"kind":{"type":"string","enum":["from","join"],"title":"Kind"},"editable":{"type":"boolean","title":"Editable"},"read_only_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Read Only Reason"},"span":{"$ref":"#/components/schemas/SqlSpanView"},"join_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Join Id"}},"type":"object","required":["id","expression","kind","editable","span"],"title":"SqlSourceView"},"SqlSpanView":{"properties":{"start":{"type":"integer","title":"Start"},"end":{"type":"integer","title":"End"}},"type":"object","required":["start","end"],"title":"SqlSpanView"},"StepView":{"properties":{"id":{"type":"string","title":"Id"},"node_kind":{"type":"string","const":"step","title":"Node Kind","default":"step"},"function_name":{"type":"string","title":"Function Name"},"block_index":{"type":"integer","title":"Block Index"},"source_span":{"$ref":"#/components/schemas/SourceSpanView"},"functional_kind":{"type":"string","title":"Functional Kind"},"display_label":{"type":"string","title":"Display Label"},"description":{"type":"string","title":"Description"},"parameters":{"items":{"$ref":"#/components/schemas/ParameterView"},"type":"array","title":"Parameters"},"csv_inputs":{"items":{"type":"string"},"type":"array","title":"Csv Inputs"},"csv_outputs":{"items":{"type":"string"},"type":"array","title":"Csv Outputs"},"parent_scope_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Scope Id"},"branch":{"anyOf":[{"type":"string","enum":["true","false"]},{"type":"null"}],"title":"Branch"},"validation_state":{"type":"string","enum":["valid","warning","unsupported"],"title":"Validation State","default":"valid"},"raw_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Code"},"read_only":{"type":"boolean","title":"Read Only","default":true},"utility":{"$ref":"#/components/schemas/UtilityView"},"capabilities":{"items":{"type":"string"},"type":"array","title":"Capabilities"}},"type":"object","required":["id","function_name","block_index","source_span","functional_kind","display_label","description","utility"],"title":"StepView"},"UtilityView":{"properties":{"name":{"type":"string","title":"Name"},"class_name":{"type":"string","title":"Class Name"},"module":{"type":"string","title":"Module"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Method"},"method_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Method Description"},"return_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Return Type"},"capabilities":{"items":{"type":"string"},"type":"array","title":"Capabilities"},"supported_mutations":{"items":{"type":"string"},"type":"array","title":"Supported Mutations"}},"type":"object","required":["name","class_name","module","title","description"],"title":"UtilityView"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"ValidationIssueView":{"properties":{"level":{"type":"string","enum":["warning","error"],"title":"Level","default":"error"},"code":{"type":"string","title":"Code"},"message":{"type":"string","title":"Message"},"parameter_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parameter Id"}},"type":"object","required":["code","message"],"title":"ValidationIssueView"},"WorkspaceDocumentRequest":{"properties":{"document_id":{"type":"string","title":"Document Id"},"source_path":{"type":"string","title":"Source Path"},"output_path":{"type":"string","title":"Output Path"},"changes":{"items":{"$ref":"#/components/schemas/ParameterChangeRequest"},"type":"array","title":"Changes"}},"type":"object","required":["document_id","source_path","output_path"],"title":"WorkspaceDocumentRequest"},"WorkspaceFileView":{"properties":{"path":{"type":"string","title":"Path"},"size_bytes":{"type":"integer","title":"Size Bytes"},"modified_at":{"type":"number","title":"Modified At"},"role":{"type":"string","enum":["source","data","generated"],"title":"Role"},"translatable":{"type":"boolean","title":"Translatable"}},"type":"object","required":["path","size_bytes","modified_at","role","translatable"],"title":"WorkspaceFileView"},"WorkspaceProjectionRequest":{"properties":{"documents":{"items":{"$ref":"#/components/schemas/WorkspaceDocumentRequest"},"type":"array","title":"Documents"}},"type":"object","title":"WorkspaceProjectionRequest"},"WorkspaceProjectionView":{"properties":{"documents":{"items":{"$ref":"#/components/schemas/ProjectedDocumentView"},"type":"array","title":"Documents"},"dependencies":{"items":{"$ref":"#/components/schemas/DependencyLinkView"},"type":"array","title":"Dependencies"},"issues":{"items":{"$ref":"#/components/schemas/DependencyIssueView"},"type":"array","title":"Issues"}},"type":"object","required":["documents","dependencies","issues"],"title":"WorkspaceProjectionView"},"WorkspaceUploadPolicyView":{"properties":{"allowed_upload_suffixes":{"items":{"type":"string"},"type":"array","title":"Allowed Upload Suffixes"},"max_upload_bytes":{"type":"integer","title":"Max Upload Bytes"},"max_file_count":{"type":"integer","title":"Max File Count"},"max_workspace_bytes":{"type":"integer","title":"Max Workspace Bytes"}},"type":"object","required":["allowed_upload_suffixes","max_upload_bytes","max_file_count","max_workspace_bytes"],"title":"WorkspaceUploadPolicyView"}}}}